Create FAKE-hacker-terminal-code
Sooooo how to use it? well first you need to be on Pc, Laptop, Mac, or anything you can use inspect element on, sadly no phones, but

How to use: First go on a website, right click the page, click inspect element, then when your on that screen, go to the top and click console, then execute the code that i gave you! go in settings to see that i made it if you dont think i made it!!!!